Rename tests in orchestrator to correspond to the main functionality (ie report, record, and delete).
