+++ test/TimeTracker.Library.Test/Services/Orchestration/DeleteTests.cs
@@ -1,0 +1,42 @@
+using System;
+using System.Linq;
+using System.Threading.Tasks;
+using FluentAssertions;
+using TimeTracker.Data;
+using TimeTracker.Library.Models;
+using TimeTracker.Library.Services;
+using TimeTracker.Library.Services.Orchestration;
+using Xunit;
+
+namespace TimeTracker.Library.Test.Services.Orchestration
+{
+    public class DeleteTests
+    {
+        private readonly TimeTrackerDbContext database;
+        private readonly SlackMessageOrchestrator orchestrator;
+
+        public DeleteTests()
+        {
+            database = new InMemoryDatabaseWithProjectsAndUsers().Database;
+            orchestrator = new SlackMessageOrchestrator(database);
+        }
+
+        [Fact]
+        public async Task HandleCommand_deleteHours_returnsDeletedMessage_andDeletesHours()
+        {
+            var user = database.Users.First();
+            var timeEntryService = new TimeEntryService(user.UserId, database);
+            await timeEntryService.CreateBillableTimeEntry(DateTime.UtcNow.Date, 7, 1, 1);
+
+            var slackMessage = await orchestrator.HandleCommand(new SlashCommandPayload()
+            {
+                text = "delete",
+                user_id = user.SlackUserId,
+                user_name = user.UserName
+            });
+
+            slackMessage.Text.Should().Be($"Deleted {7d:F1} hours for date: {DateTime.UtcNow.Date:D}");
+            database.TimeEntries.Count().Should().Be(0);
+        }
+    }
+}
